+++ packages/core/src/store/sagas.ts
@@ -1,8 +1,9 @@
 import { put, takeEvery, all } from 'redux-saga/effects'
-import { ActionType, ChangeAction } from './reducer'
+import { ActionType, AsyncChangeAction } from './actions'
 
-export function* asyncChangeAction(action: ChangeAction) {
-  yield put({ type: ActionType.Change, payload: action.payload })
+export function* asyncChangeAction(action: AsyncChangeAction) {
+  const payload = yield action.payload
+  yield put({ type: ActionType.Change, payload: payload })
 }
 
 export function* watchAction() {
