+++ packages/core/src/store/selectors.ts
@@ -1,0 +1,69 @@
+import { Plugin, State } from '@edtr-io/core'
+import { BaseState } from './reducer'
+import { PluginState, PluginType } from '../plugin'
+
+/** Selectors */
+export function getRoot(state: State) {
+  return state.root
+}
+
+export function getDocuments(state: State): Record<string, PluginState> {
+  return state.documents
+}
+
+export function getDocument(state: State, id: string): PluginState | null {
+  return getDocuments(state)[id] || null
+}
+
+export function getPlugin(state: State, type: string): Plugin | null {
+  const plugins = getPlugins(state)
+
+  return plugins[type] || null
+}
+
+export function getClipboard(state: State): PluginState[] {
+  return state.clipboard
+}
+
+export function getPluginOrDefault(
+  state: State,
+  type = getDefaultPlugin(state)
+): Plugin | null {
+  return getPlugin(state, type)
+}
+
+export function getDefaultPlugin(state: State): PluginType {
+  return state.defaultPlugin
+}
+
+export function getPluginTypeOrDefault(
+  state: State,
+  type = getDefaultPlugin(state)
+): PluginType {
+  return type
+}
+
+export function getPlugins<K extends string = string>(
+  state: State
+): Record<K, Plugin> {
+  return state.plugins
+}
+
+export function isFocused(state: State, id: string): boolean {
+  return state.focus === id
+}
+
+export function isEditable(state: State): boolean {
+  return state.editable
+}
+
+export function hasPendingChanges(state: State): boolean {
+  return state.history.pending !== 0
+}
+export function pendingChanges(state: State): number {
+  return state.history.pending
+}
+
+export function hasHistory(state: BaseState | State): state is State {
+  return (state as State).history !== undefined
+}
